% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circularEV-package.R
\docType{package}
\name{circularEV-package}
\alias{circularEV}
\alias{circularEV-package}
\title{circularEV: Extreme Value Analysis for Circular Data}
\description{
General functions for performing extreme value analysis on a circular 
    domain as part of the statistical methodology in the paper by Konzen, E., 
    Neves, C., and Jonathan, P. (2021). Modeling nonstationary extremes of storm severity: Comparing parametric and semiparametric inference. Environmetrics, 32(4), e2667.
}
\keyword{internal}
