% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drc.R
\docType{data}
\name{drc}
\alias{drc}
\title{Directional covariate for HsSP data}
\format{A vector of length 1521.}
\usage{
drc
}
\description{
A vector of length 1521.
}
\details{
Directional covariate of HsSP data used by Reistad, M., Breivik, Ø.,
  Haakenstad, H., Aarnes, O. J., Furevik, B. R., and Bidlot, J.-R. (2011), A
  high-resolution hindcast of wind and waves for the North Sea, the Norwegian
  Sea, and the Barents Sea, J. Geophys. Res., 116:1-18.
}
\keyword{datasets}
