% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HsSP.R
\docType{data}
\name{HsSP}
\alias{HsSP}
\title{Hindcast storm peak significant wave height data}
\format{A vector of length 1521.}
\usage{
HsSP
}
\description{
A vector of length 1521.
}
\details{
HsSP data used by Reistad, M., Breivik, Ø., Haakenstad, H., Aarnes,
  O. J., Furevik, B. R., and Bidlot, J.-R. (2011), A high-resolution hindcast
  of wind and waves for the North Sea, the Norwegian Sea, and the Barents
  Sea, J. Geophys. Res., 116:1-18.
}
\keyword{datasets}
