% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotParamEstim.R
\name{PlotParamEstim}
\alias{PlotParamEstim}
\title{Plot of parameter estimates with bootstrap confidence intervals}
\usage{
PlotParamEstim(
  bootEstimates,
  thetaGrid = 1:360,
  alpha = 0.05,
  ylim = NULL,
  cex.axis = 15,
  cex.lab = 2,
  thrWidth = 2,
  ylab = NULL,
  thrColor = "#D45E1A"
)
}
\arguments{
\item{bootEstimates}{Bootstrap estimates (for example, shape or scale)}

\item{thetaGrid}{Grid values at which the estimation is performed}

\item{alpha}{Significance level for the confidence intervals. Default to 0.05.}

\item{ylim}{Range for the y-axis}

\item{cex.axis}{Graphical parameter}

\item{cex.lab}{Graphical parameter}

\item{thrWidth}{Threshold width}

\item{ylab}{y-axis label}

\item{thrColor}{Threshold colour}
}
\value{
Plot of parameter estimates.
}
\description{
Plot of parameter estimates with bootstrap confidence intervals
}
\examples{
## See examples in vignettes:
# vignette("localMethods", package = "circularEV")
# vignette("splineML", package = "circularEV")
}
\seealso{
\code{\link{SplineML}} and \code{\link{LocalEstim}} for examples.
}
