% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circularEV-package.R
\docType{package}
\name{circularEV-package}
\alias{circularEV}
\alias{circularEV-package}
\title{circularEV: Extreme Value Analysis for Circular Data}
\description{
General functions for performing extreme value analysis on a circular 
    domain as part of the statistical methodology in the paper by Konzen, E., 
    Neves, C., and Jonathan, P. (2020+). Modelling non-stationary extremes of 
    storm severity: comparing parametric and semi-parametric inference. 
    Environmetrics (to appear).
}
\keyword{internal}
